+++ src/utils/isEmptyArray.js
@@ -1,0 +1,8 @@
+const isNonEmptyArray = require('./isNonEmptyArray');
+
+/**
+ * Check if a value is a empty `Array`.
+ */
+const isEmptyArray = (value) => !isNonEmptyArray(value);
+
+module.exports = isEmptyArray;