+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Fixed (commit 3)
+
+- Added a new endpoint, `/contracts`, which returns a list of contracts belonging to a user (client or contractor).
+- The list should only contain non-terminated contracts.
+
+
 ### Fixed (commit 2)
 
 - Fixed an issue where the user was able to see contracts that did not belong to them. The code has been modified to ensure that the user can only view their own contracts.